Add Xcode 9 support
With the large number of simulator changes in Xcode 9 came a change
where the notification we're hijacking needs to send along the UDIDs of
the simulators where the change should take effect. This is because you
can have multiple running simulators with different locations.

To do this, for now, we're just reading the output of `simctl list` and
sending all booted UDIDs along.