Reorder failure errors
When there are no locations found for a query, MapKit returns a NSError
instead of an empty array. This resulted in us printing a opaque error.
Now we'll do that last, only if there is an error, but we also found a
place.

Here's an example of a query that currently fails:

Lyft HQ San Francisco